Evaluation System For Teachers and Students for acedemic use.
This is Online Evaluation System for Academic use. Teachers and Students can use this system. Teachers can upload the assignments , Solutions and Study Resources to the system. Students who have enrolled for the course get notified once resource is avaliable. Teachers can evaluate the assignments and upload the marks and then students can find them on their login. Appointment of Teachers is done by admin. Students who wish to enroll for course need to seek the permission for either Admin or Teachers.